chore(core): move `Storage` class to utils dir
