+++ packages/core/src/composables/useVueFlow.ts
@@ -1,7 +1,7 @@
 import type { EffectScope } from 'vue'
 import type { FlowProps, VueFlowStore } from '~/types'
 import { VueFlow } from '~/context'
-import { Storage } from '~/storage'
+import { Storage } from '~/utils'
 
 type Injection = VueFlowStore | null | undefined
 type Scope = (EffectScope & { vueFlowId: string }) | undefined
