+++ packages/core/src/utils/index.ts
@@ -2,4 +2,5 @@ export * from './edge'
 export * from './graph'
 export * from './node'
 export * from './changes'
+export * from './storage'
 export * from './store'
